Move filtering results by blast facets functions into utils.
